Move some namespaces on Factory
