+++ docs/resource_factories.md
@@ -27,8 +27,7 @@ declare(strict_types=1);
 
 namespace App\Factory;
 
-use App\Entity\Book;
-use Sylius\Component\Resource\Factory\FactoryInterface;
+use App\Entity\Book;use Sylius\Resource\Factory\FactoryInterface;
 
 final class BookFactory implements FactoryInterface
 {
@@ -60,10 +59,7 @@ declare(strict_types=1);
 
 namespace App\Factory;
 
-use App\Entity\Book;
-use Symfony\Component\Security\Core\Security;
-use Sylius\Component\Resource\Factory\FactoryInterface;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
+use App\Entity\Book;use Sylius\Resource\Factory\FactoryInterface;use Symfony\Component\Security\Core\Security;
 
 final class BookFactory implements FactoryInterface
 {
@@ -129,9 +125,7 @@ declare(strict_types=1);
 
 namespace App\Factory;
 
-use App\Entity\Book;
-use Sylius\Component\Resource\Factory\FactoryInterface;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
+use App\Entity\Book;use Sylius\Component\Resource\Repository\RepositoryInterface;use Sylius\Resource\Factory\FactoryInterface;
 
 final class BookFactory implements FactoryInterface
 {
