+++ src/Component/Factory/TranslatableFactory.php
@@ -13,36 +13,10 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Resource\Factory;
 
-use Sylius\Component\Resource\Exception\UnexpectedTypeException;
-use Sylius\Component\Resource\Model\TranslatableInterface;
-use Sylius\Component\Resource\Translation\Provider\TranslationLocaleProviderInterface;
+class_exists(\Sylius\Resource\Factory\TranslatableFactory::class);
 
-final class TranslatableFactory implements TranslatableFactoryInterface
-{
-    private FactoryInterface $factory;
-
-    private TranslationLocaleProviderInterface $localeProvider;
-
-    public function __construct(FactoryInterface $factory, TranslationLocaleProviderInterface $localeProvider)
+if (false) {
+    final class TranslatableFactory extends \Sylius\Resource\Factory\TranslatableFactory
     {
-        $this->factory = $factory;
-        $this->localeProvider = $localeProvider;
-    }
-
-    /**
-     * @throws UnexpectedTypeException
-     */
-    public function createNew()
-    {
-        $resource = $this->factory->createNew();
-
-        if (!$resource instanceof TranslatableInterface) {
-            throw new UnexpectedTypeException($resource, TranslatableInterface::class);
-        }
-
-        $resource->setCurrentLocale($this->localeProvider->getDefaultLocaleCode());
-        $resource->setFallbackLocale($this->localeProvider->getDefaultLocaleCode());
-
-        return $resource;
     }
 }
