+++ src/Component/spec/Factory/TranslatableFactorySpec.php
@@ -15,10 +15,10 @@ namespace spec\Sylius\Component\Resource\Factory;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Exception\UnexpectedTypeException;
-use Sylius\Component\Resource\Factory\FactoryInterface;
-use Sylius\Component\Resource\Factory\TranslatableFactoryInterface;
 use Sylius\Component\Resource\Model\TranslatableInterface;
 use Sylius\Component\Resource\Translation\Provider\TranslationLocaleProviderInterface;
+use Sylius\Resource\Factory\FactoryInterface;
+use Sylius\Resource\Factory\TranslatableFactoryInterface;
 
 final class TranslatableFactorySpec extends ObjectBehavior
 {
