+++ src/Component/src/Factory/FactoryInterface.php
@@ -11,15 +11,17 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Factory;
+namespace Sylius\Resource\Factory;
 
-class_exists(\Sylius\Resource\Factory\Factory::class);
-
-if (false) {
+/**
+ * @template T of object
+ */
+interface FactoryInterface
+{
     /**
-     * @inheritDoc
+     * @return T
      */
-    final class Factory extends \Sylius\Resource\Factory\Factory
-    {
-    }
+    public function createNew();
 }
+
+class_alias(FactoryInterface::class, \Sylius\Component\Resource\Factory\FactoryInterface::class);
