+++ src/Component/src/State/Factory.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 namespace Sylius\Resource\State;
 
 use Psr\Container\ContainerInterface;
-use Sylius\Component\Resource\Factory\FactoryInterface as ResourceFactoryInterface;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\Factory\FactoryInterface as ResourceFactoryInterface;
 use Sylius\Resource\Metadata\FactoryAwareOperationInterface;
 use Sylius\Resource\Metadata\Operation;
 use Sylius\Resource\Symfony\ExpressionLanguage\ArgumentParserInterface;
