+++ tests/Application/src/Subscription/Factory/SubscriptionFactory.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace App\Subscription\Factory;
 
 use App\Subscription\Entity\Subscription;
-use Sylius\Component\Resource\Factory\FactoryInterface;
+use Sylius\Resource\Factory\FactoryInterface;
 
 final class SubscriptionFactory implements FactoryInterface
 {
